+++ app/components/freefeed_user_component.rb
@@ -8,4 +8,8 @@ class FreefeedUserComponent < ViewComponent::Base
   def username
     @user["username"] || @user[:username]
   end
+
+  def userpic_url
+    @user["profile_picture_url"] || @user[:profile_picture_url]
+  end
 end